file moving updates
several functions moved from init.py , indentation error updates
